POC of quick navigation (#20)
With help of the plugin `rehype-slug` we now add ID's to all the headings. Those id's we can then traverse in a `useEffect` to create a "table of contents" navigation aka Quick nav.

Whenever user scrolls it will try to find out which headline is in focus marked by a underline for now.

No UI is added, will come later.