Quick nav (#21)
This PR introduces the Quick nav `rehype-autolink-headings` we now have ID's on the headings we can link to. 
The UI is added, but we might need to tweek the observar margin. For now it's set to `-5% 0px -50% 0px` in the `QuickNav.tsx` file.
I also added the package `rehype-sanitize` for better security.