Icon refactoring (#36)
This PR adds the correct sizes to all icons and fixed the typing issues.

@dvdslns I added the specific sizes on some of the icons you added. Lauren will properly change the icons so they are all 16x16 or something similar. We just need them all to be same size. 

I removed the `'aria-hidden': false,` prop since it should already be false if not defined. I also removed the `focusable: 'auto'` in previous PR. Im not sure if we event need this prop, i think its outdated, but please let me know if it's needed anyway?